+++ src/editor.rs
@@ -4,8 +4,8 @@ mod validation;
 
 use crate::editor::highlighter::Highlighter;
 use crate::preferences::Preferences;
-use crate::widget::{Element, FragmentShader};
-use crate::{preferences, theme, JETBRAINS_MONO};
+use crate::widget::Element;
+use crate::{preferences, theme, FragmentShader, JETBRAINS_MONO};
 use iced::alignment::Horizontal;
 use iced::widget::{button, checkbox, column, container, row, text, text_editor, tooltip};
 use iced::{alignment, keyboard, Alignment, Command, Font, Length};
